+++ src/QueryType/ManagedResources/Query/Command/Remove.php
@@ -37,24 +37,4 @@ class Remove extends AbstractCommand
     {
         return Request::METHOD_DELETE;
     }
-
-    /**
-     * Empty.
-     *
-     * @return string
-     */
-    public function getRawData(): string
-    {
-        return '';
-    }
-
-    /**
-     * Empty.
-     *
-     * @return string
-     */
-    public function getTerm(): string
-    {
-        return '';
-    }
 }