+++ tests/Core/Query/ResponseParserTest.php
@@ -3,7 +3,14 @@
 namespace Solarium\Tests\Core\Query;
 
 use PHPUnit\Framework\TestCase;
+use Solarium\Core\Client\Response;
+use Solarium\Core\Query\AbstractQuery;
 use Solarium\Core\Query\AbstractResponseParser;
+use Solarium\Core\Query\RequestBuilderInterface;
+use Solarium\Core\Query\ResponseParserInterface;
+use Solarium\Core\Query\Result\Result;
+use Solarium\Core\Query\Result\ResultInterface;
+use Solarium\Core\Query\Status4xxNoExceptionInterface;
 
 class ResponseParserTest extends TestCase
 {
@@ -40,6 +47,28 @@ class ResponseParserTest extends TestCase
         );
     }
 
+    /**
+     * @testWith [200, true]
+     *           [400, false]
+     */
+    public function testParseStatus(int $statusCode, bool $expectedSuccess)
+    {
+        $result = new Result(
+            new TestQueryForResponseParser(),
+            new Response('', [sprintf('HTTP/1.1 %d Status', $statusCode)])
+        );
+
+        $expected = [
+            'wasSuccessful' => $expectedSuccess,
+            'statusMessage' => 'Status',
+        ];
+
+        $this->assertSame(
+            $expected,
+            $this->parser->parse($result)
+        );
+    }
+
     public function testAddHeaderInfo()
     {
         $data = [
@@ -72,12 +101,31 @@ class ResponseParserTest extends TestCase
     }
 }
 
+class TestQueryForResponseParser extends AbstractQuery implements Status4xxNoExceptionInterface
+{
+    public function getType(): string
+    {
+        return 'testType';
+    }
+
+    public function getRequestBuilder(): RequestBuilderInterface
+    {
+        return null;
+    }
+
+    public function getResponseParser(): ResponseParserInterface
+    {
+        return null;
+    }
+}
+
 /**
  * Dummy implementation to test code in abstract class.
  */
 class TestResponseParser extends AbstractResponseParser
 {
-    public function parse($result)
+    public function parse(ResultInterface $result): array
     {
+        return $this->parseStatus([], $result);
     }
 }