+++ CHANGELOG.md
@@ -5,6 +5,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [unreleased]
+### Added
+- Solarium\QueryType\ManagedResources\Result\Command::getWasSuccessful()
+- Solarium\QueryType\ManagedResources\Result\Command::getStatusMessage()
+- Query a single term in a Managed Resource
+
+### Fixed
+- Syntax error in request with facet queries that contain local parameters
+- HEAD requests could lead to timeouts with cURL adapter
+
+### Changed
+
+### Removed
+
+### Deprecated
+
+
 ## [6.1.3]
 ### Fixed
 - possible exception in Debug\Detail::__toString() when sub details are missing